Cast length to long before putting into metadata.  Fixes #22
https://github.com/couchbase/couchbase-lite-android-core/issues/22

+++ src/main/java/com/couchbase/lite/Attachment.java
@@ -182,7 +182,7 @@ public class Attachment {
                 if (body != null) {
                     // Copy attachment body into the database's blob store:
                     BlobStoreWriter writer = blobStoreWriterForBody(body, database);
-                    metadataMutable.put("length", writer.getLength());
+                    metadataMutable.put("length", (long)writer.getLength());
                     metadataMutable.put("digest", writer.mD5DigestString());
                     metadataMutable.put("follows", true);
                     database.rememberAttachmentWriter(writer);
